requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0

# LLM and AI
google-generativeai>=0.3.0
python-dotenv>=0.19.0
transformers>=4.30.0
deep-translator>=1.11.0

# Date and time processing
python-dateutil>=2.8.0

# Web scraping
undetected-chromedriver>=3.5.0
selenium>=4.0.0
beautifulsoup4>=4.11.0
webdriver-manager>=3.8.0

# Database
pymysql>=1.0.0
sqlalchemy>=1.4.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
requests>=2.28.0
